refactor: merge overlay service with calendars service
