+++ apps/api/v2/src/ee/calendars/calendars.module.ts
@@ -1,6 +1,5 @@
 import { CalendarsController } from "@/ee/calendars/controllers/calendars.controller";
 import { CalendarsService } from "@/ee/calendars/services/calendars.service";
-import { OverlayCalendarsService } from "@/ee/calendars/services/overlay-calendars.service";
 import { CredentialsRepository } from "@/modules/credentials/credentials.repository";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
 import { UsersModule } from "@/modules/users/users.module";
@@ -8,8 +7,8 @@ import { Module } from "@nestjs/common";
 
 @Module({
   imports: [PrismaModule, UsersModule],
-  providers: [CredentialsRepository, CalendarsService, OverlayCalendarsService],
+  providers: [CredentialsRepository, CalendarsService],
   controllers: [CalendarsController],
-  exports: [OverlayCalendarsService],
+  exports: [CalendarsService],
 })
 export class CalendarsModule {}
